+++ game/client/events/mouse_scroll.hpp
@@ -13,7 +13,7 @@
 namespace events
 {
 struct MouseScroll final {
-    float dx;
-    float dy;
+    double dx;
+    double dy;
 };
 } // namespace events
